Refactor mellanox_thermal_control_test_helper.py
